requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
pypdf
faiss-cpu
langchainhub
langsmith
langgraph
langchain-pinecone
requests
black
openai
google-search-results
python-dotenv
beautifulsoup4